% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify5HT_helpers.R
\name{assert_has_cols}
\alias{assert_has_cols}
\title{Assert that a data.frame has required columns}
\usage{
assert_has_cols(df, cols, df_name = deparse(substitute(df)))
}
\arguments{
\item{df}{data.frame}

\item{cols}{character vector of required column names}

\item{df_name}{label used in error message}
}
\description{
Assert that a data.frame has required columns
}
\keyword{internal}
