% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_metrics5HT.R
\name{.fano}
\alias{.fano}
\title{Fano factor of a rate sample (variance / mean), guarded}
\usage{
.fano(x)
}
\description{
Fano factor of a rate sample (variance / mean), guarded
}
\keyword{internal}
