% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample_trace_adaptive.R
\name{downsample_trace_adaptive}
\alias{downsample_trace_adaptive}
\title{Adaptive Downsampling}
\usage{
downsample_trace_adaptive(
  v,
  sr,
  dvdt_thresh_mV_ms = 5,
  slow_bin_ms = 10,
  fast_keep_every = 1,
  max_points = 50000,
  thin_slow_every = 2
)
}
\description{
Adaptive Downsampling
}
