% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify5HT_helpers.R
\name{find_runs}
\alias{find_runs}
\title{Find contiguous TRUE runs in a logical vector (with time positions)}
\usage{
find_runs(time, flag)
}
\arguments{
\item{time}{numeric time axis}

\item{flag}{logical vector (same length as time)}
}
\value{
tibble with start, end, dur for each TRUE run
}
\description{
Find contiguous TRUE runs in a logical vector (with time positions)
}
\keyword{internal}
