% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify5HT_helpers.R
\name{first_non_missing}
\alias{first_non_missing}
\title{Return the first non-missing entry in a vector}
\usage{
first_non_missing(x, default = NA)
}
\description{
Return the first non-missing entry in a vector
}
\keyword{internal}
