% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify5HT_helpers.R
\name{get_extreme_in_window}
\alias{get_extreme_in_window}
\title{Get the extreme (min or max) value inside a time window}
\usage{
get_extreme_in_window(time, y, window, which = c("min", "max"))
}
\description{
Get the extreme (min or max) value inside a time window
}
\keyword{internal}
