% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_classify5HT.R
\name{plot_centered_violin}
\alias{plot_centered_violin}
\title{Violin + jitter of response peak by subclass}
\usage{
plot_centered_violin(per_cell)
}
\arguments{
\item{per_cell}{one-row-per-cell tibble.}
}
\value{
ggplot object.
}
\description{
Violin + jitter of response peak by subclass
}
