% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_classify5HT.R
\name{plot_response_composition}
\alias{plot_response_composition}
\title{Stacked bar:  response composition by subclass}
\usage{
plot_response_composition(per_cell, x = "assigned_subclass")
}
\arguments{
\item{per_cell}{one-row-per-cell tibble from \code{build_per_cell_classifier()}.}

\item{x}{column name to group on (default "assigned_subclass").}
}
\value{
ggplot object.
}
\description{
Stacked bar:  response composition by subclass
}
