% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_classify5HT.R
\name{plot_response_heatmap}
\alias{plot_response_heatmap}
\title{Tile heatmap:  response call vs subclass}
\usage{
plot_response_heatmap(per_cell, x = "response_call", y = "assigned_subclass")
}
\arguments{
\item{per_cell}{one-row-per-cell tibble.}

\item{x, y}{column names for tile axes.}
}
\value{
ggplot object.
}
\description{
Tile heatmap:  response call vs subclass
}
