% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_rmp_protocol_set.R
\name{plot_rmp_protocol}
\alias{plot_rmp_protocol}
\title{Plot an RMP protocol object (trace or features)}
\usage{
plot_rmp_protocol(rmp_protocol, what = c("trace", "features"))
}
\arguments{
\item{rmp_protocol}{One element of \code{srt$dfs$rmp$by_protocol}.}

\item{what}{"trace" or "features".}
}
\value{
A ggplot object (or NULL).
}
\description{
Plot an RMP protocol object (trace or features)
}
