% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify5HT_helpers.R
\name{response_cols}
\alias{response_cols}
\title{Standard colour palette for response classes}
\usage{
response_cols
}
\description{
Standard colour palette for response classes
}
\keyword{internal}
