% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spikePuff5HT.R
\name{spikePuff5HT}
\alias{spikePuff5HT}
\title{Spike puff analysis to be nnested in the srt nnest}
\usage{
spikePuff5HT(x)
}
\arguments{
\item{x}{5ht nnest}
}
\value{
srt nnest
}
\description{
Spike puff analysis to be nnested in the srt nnest
}
\examples{
x = i
x = srt


}
